将每个测试方法改为相互独立，原因是： 1.  相互独立的单元测试，有助于debug问题， 2.  不需要考虑执行顺序，额外的运行环境等，更加简单。 做的工作是： 1. 去掉了顺序执行,这有助于并发调用，减少测试时间 2. 去掉了测试方法中共享的实例

+++ Server/ServerTest/FilesSeed.cs
@@ -184,8 +184,14 @@ public class FilesSeed : IDisposable
 
     public void Dispose()
     {
-        Directory.Delete($"{TestPath}/OldDir", true);
-        Directory.Delete($"{TestPath}/NewDir", true);
+        if (Directory.Exists($"{TestPath}/OldDir"))
+        {
+            Directory.Delete($"{TestPath}/OldDir", true);
+        }
+        if (Directory.Exists($"{TestPath}/NewDir"))
+        {
+            Directory.Delete($"{TestPath}/NewDir", true);
+        }
         Console.WriteLine("FilesSeed Dispose");
         GC.SuppressFinalize(this);
     }
